requirements.txt
black==20.8b1
celery==4.4.2
vine==1.3.0
django==3.1.3
djangorestframework==3.11.0
numpy==1.19.4
pandas==1.1.4
psycopg2==2.8.6
redis==3.4.1
names==0.3.0